% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readText.R
\name{readText}
\alias{readText}
\title{readText}
\usage{
readText(folderPath)
}
\arguments{
\item{folderPath}{A file path to a directory of .txt file containing text data.}
}
\value{
Returns a data frame concatenating the text data of each file and using file names as document IDs.

myfunction(1) # returns 1
}
\description{
\code{myfunction} reads and formats a directory of plaintext files.
}
