% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformText.R
\name{transformText}
\alias{transformText}
\title{transformText}
\usage{
transformText(textData, returnFullData = F)
}
\arguments{
\item{textData}{Data frame with at least two rows: ID, a unique ID for each text sample and Text, the text samples. Column names must match exactly.}

\item{returnFullData}{Should result data frame report individual word ratings (T) or averaged by sample (F)}
}
\value{
A dataframe with ratings for 22 emotional dimensions bound to each text sample.
}
\description{
\code{transformText} aligns each text sample with associated values on 22 emotional dimensions
}
